@@ -30,17 +30,21 @@ async function upsertNationality(connection, nationality, nationalitylogo) {
 
 // 更新player信息
 async function updatePlayer(connection, playerData) {
-    const query = 'INSERT INTO players (vlrid, gameid, realname, nationality, teamid, upd_time) VALUES (?, ?, ?, ?, ?, NOW()) ON DUPLICATE KEY UPDATE gameid = ?, realname = ?, nationality = ?, teamid = ?, upd_time = NOW()';
+    const query = 'INSERT INTO players (vlrid, gameid, realname, birthdate, nationality, teamid, profile_url, upd_time) VALUES (?, ?, ?, ?, ?, ?, ?, NOW()) ON DUPLICATE KEY UPDATE gameid = ?, realname = ?, birdate = ?, nationality = ?, teamid = ?, profile_url = ?, upd_time = NOW()';
     await connection.execute(query, [
         playerData.info.id,
         playerData.info.user,
         playerData.info.name,
+        playerData.birthdate,
         playerData.info.country,
         playerData.team.id,
+        playerData.info.img,
         playerData.info.user,
         playerData.info.name,
+        playerData.birthdate,
         playerData.info.country,
-        playerData.team.id
+        playerData.team.id,
+        playerData.info.img
     ]);
 }
 
@@ -59,6 +63,43 @@ async function fetchPlayerFromAPI(vlrid) {
     }
 }
 
+async function fetchBirthFromLiquid(gameid) {
+    console.log(`Fetching birth from liquipedia ${gameid}...`);
+    const response = await axios.get(`https://liquipedia.net/valorant/api.php`, {
+        params: {
+            action: 'query',
+            format: 'json',
+            prop: 'extracts',
+            titles: gameid,
+            exintro: 1,
+            explaintext: 1
+        },
+        headers: {
+            'User-Agent': 'ValorantPlayerLookup/1.0',
+            'Accept-Encoding': 'gzip'
+        }
+    });
+
+    const pages = response.data.query.pages;
+    const pageId = Object.keys(pages)[0];
+    
+    if (pageId === '-1') return null;
+    
+    const content = pages[pageId].extract;
+    const regex = /born\s+([A-Za-z]+)\s+(\d{1,2}),\s*(\d{4})/i;
+    const dobMatch = content.match(regex);
+    const monthMap = {
+        'january': '01', 'february': '02', 'march': '03', 'april': '04',
+        'may': '05', 'june': '06', 'july': '07', 'august': '08',
+        'september': '09', 'october': '10', 'november': '11', 'december': '12'
+    };
+    const birthdate = dobMatch 
+        ? `${dobMatch[3]}-${monthMap[dobMatch[1].toLowerCase()]}-${dobMatch[2].padStart(2, '0')}`
+        : null;
+    console.log(`Fetched birth from liquipedia ${gameid}: ${birthdate}`);
+    return birthdate;
+}
+
 // 更新useagents信息
 async function updatePlayerAgents(connection, vlrid, agents) {
     // 删除现有的agent记录
@@ -75,7 +116,7 @@ async function updatePlayerAgents(connection, vlrid, agents) {
     }
 }
 
-async function updateFromAPI(connection, vlrid) {
+async function updateFromAPI(connection, vlrid, gameid) {
     const MAX_RETRIES = 3;
     let retryCount = 0;
     
@@ -86,7 +127,12 @@ async function updateFromAPI(connection, vlrid) {
         
         console.log(`Fetching data from API ${vlrid}...`);
         const apiData = await fetchPlayerFromAPI(vlrid);
-        let playerData = apiData.data;
+        const birthdate = await fetchBirthFromLiquid(gameid);
+
+        let playerData = {
+            ...apiData.data,
+            birthdate: birthdate
+        };
 
         await connection.beginTransaction();
         await upsertTeam(connection, playerData.team);
@@ -95,22 +141,29 @@ async function updateFromAPI(connection, vlrid) {
         await updatePlayerAgents(connection, playerData.info.id, playerData.agents);
         await connection.commit();
 
-        return {
-            "vlrid": playerData.info.id,
-            "gameid": playerData.info.user,
-            "realname": playerData.info.name,
-            "nationality": playerData.info.country,
-            "teamname": playerData.team.name,
-            "teamlogo": playerData.team.logo,
-            "agents": playerData.agents.slice(0, 3).map(agent => ({
-                "agent": agent.agentName,
-                "roundsPlayed": agent.roundsPlayed
-            }))
-        };
+        // return {
+        //     "vlrid": playerData.info.id,
+        //     "gameid": playerData.info.user,
+        //     "realname": playerData.info.name,
+        //     "nationality": playerData.info.country,
+        //     "teamname": playerData.team.name,
+        //     "teamlogo": playerData.team.logo,
+        //     "agents": playerData.agents.slice(0, 3).map(agent => ({
+        //         "agent": agent.agentName,
+        //         "roundsPlayed": agent.roundsPlayed
+        //     }))
+        // };
     } catch (error) {retryCount++;}}
+}
+
+async function updatebirthFromAPI(connection, gameid) {
+    const birthdate = await fetchBirthFromLiquid(gameid);
+    const query = `UPDATE players SET birthdate =${birthdate? ('\'' + birthdate + '\''):  null} WHERE gameid ='${gameid}'`;
+    await connection.execute(query);
 }
 
 module.exports = {
    isDataExpired,
-   updateFromAPI
+   updateFromAPI,
+   updatebirthFromAPI
 }
